requirements.txt
pytest==7.3.1
webdriver-manager==3.8.6
pytest-html==3.2.0
pytest-xdist==3.3.1
openpyxl==3.1.2
allure-pytest==2.13.2
selenium==4.9.1

